Use pointer cursor
For so long it has bugged me that custom style input elements often have pointer cursors, as if they were links, and not the default cursor as inputs natively have. Now I've however gotten so used to it thiugh that I think it feels odd not to have a pointer cursor on the inputs. I'm chocked!